DatePicker demo 2
Same as previous DatePicker demonstration but this example shows how to set the date prior to the Java FX stage being displayed.  In other words, you can set a date for the DatePicker to show without having to use the calendar to select a date yourself. Of course, you can always use the calendar to override the date displayed to select another date if you wish.