root changed to a BorderPane
The 'root' has now been changed to use a BorderPane layout so as to be able to incorporate a simple button control on the form as well as a tree. The botton has a 'setOnAction' event handler using a lambda expression which is invoked whenever the button is clicked with the mouse. When this happens, a message is written to the console and the program exits.